Update main.cpp
updated type of variables

+++ main.cpp
@@ -1,16 +1,16 @@
 #include <stdio.h>
 #include <math.h>
 
-float findDiscriminant(int a, int b, int c)
+double findDiscriminant(int a, int b, int c)
 {
   return b * b - 4 * a * c; 
 }
 
 float* input()
 {
-  float a = NAN;
-  float b = NAN;
-  float c = NAN;
+  double a = NAN;
+  double b = NAN;
+  double c = NAN;
 
   printf("Введите коэфиценты уравнения a*x^2 + b*x + c = 0\n");
   scanf("%f %f %f", &a, &b, &c); 
@@ -24,7 +24,7 @@ float* input()
   return arg; 
 }
 
-void linealEquation(float b, float c)
+void linealEquation(double b, double c)
 {
     if (b == 0)
     {
@@ -37,9 +37,9 @@ void linealEquation(float b, float c)
       printf("Единственный корень: x = %.3f\n", -c / b);  
 }
 
-void quadraticEquation(float a, float b, float c)
+void quadraticEquation(double a, double b, double c)
 {
-    float discriminant = findDiscriminant(a, b, c); 
+    double discriminant = findDiscriminant(a, b, c); 
 
     if (discriminant < 0)
     {
@@ -54,8 +54,8 @@ void quadraticEquation(float a, float b, float c)
     else 
     {
 
-    float x1 = NAN;
-    float x2 = NAN;
+    double x1 = NAN;
+    double x2 = NAN;
 
     printf("Дискриминант равен: %f\n", sqrt(discriminant));
 
@@ -68,10 +68,10 @@ void quadraticEquation(float a, float b, float c)
 
 int main()
 {
-  float a = NAN;
-  float b = NAN;
-  float c = NAN;
-  float discriminant = NAN;
+  double a = NAN;
+  double b = NAN;
+  double c = NAN;
+  double discriminant = NAN;
   float* arg = NULL;
